use-cases/generative-feedback-loop/requirements.txt
glassflow
supabase
python-dotenv
streamlit
weaviate-client==4.7.1